Many changes to existing classes and added some new (see below).
AtomicRelationDescription.java: In general this class now does the work
which was done by the RelationDescription class.
Added: getSubstitutions() to return the substitutions of this class as a
List<Pair<QuantifiableVariable, Term>>.
the createSubstitutionForFunction from RelationDescription.
varsToTerm from RelationDescription.
generateName from RelationDescription.

ConstructorExtractor.java: Now gets its function-namespace from the
Services not from the Proof.

RelationDescription.java: In general it passes most of its work to the
class AtomicRelationDescription (see above).

RelationDescriptionDialog.java: Now displays a maximum of 10
AtomicRelationDescriptions for each RelationDescription (to avoid out of
memory errors).

SortExtractor.java: Now also extracts the sorts from terms (not only
variables).

RelationDescriptionFactory.java: This is a new class with generates all
RelationDescriptions for a Term.

InductionFormulaCreator.java: Creates the formula for the Taclet to
replace the given term.

InductionTacletGenerator.java: This class create the
structural-induction taclet.

RelationDescriptionGenerator.java: renamed class
RelationDescriptionFactory to this name.