fixes 1504
The problem is that ":" is not valid in Windows paths (or at least in
windows paths of NIO). OriginTermLabel.java uses `Paths.get` to parse
the given filename and try to get the filename. After fix, it uses the
more liberal File API.

The main problem is that the paths encoded in the `PositionedString`
(used in the AST) are not accurate: They are neither a valid or a valid URL:

```
 /C:/Users/weigl/IdeaProjects/key/key/key.ui/URL:jar:file:/C:/Users/weigl/IdeaProjects/key/key/key.core/build/libs/key.core-2.7.0-da9d9e7adc.jar!/de/uka/ilkd/key/java/JavaRedux/java/lang/NullPointerException.java
 ```

 The first part `/C:/Users/weigl/IdeaProjects/key/key/key.ui/URL:` is not
 a valid protocol. Then the original Java URL comes up:
 `jar:file:...`

 Someone with more *deep KeY* experience should investigate this.
 It seemes to be introduced in `SLEEnvInput`.

(cherry picked from commit 4d614d995afa5ad33acead0a6c4b5feb42a7f2d7)