Merge branch 'pelevinaAbstractContracts' into 'master'
This commit merges support for abstract contracts in KeY 2.7. It was previously
implemented on the branch pelevinaAbstractContracts for KeY 2.1.
Abstract contracts allow to create partial proofs (while the concrete requires,
ensures and assignable clauses of a method contract are yet unknown), which lead
to first-order proof obligations that can be completed later on.