Fix taclet application exception
KeY tries to add formulas close to the position where a taclet has been applied. In case of a rewrite taclet without a replacewith, position information was not correctly updated as it was erroneously assumed that then no position information becomes invalidated. That is sadly not correct to term label refactoring steps still taking place.

I experienced the issue (thrown exception) with rule "onlyCreatdedObjectsAreReferenced"