do not register self in the service namespace in StateVars registering self in the namespace there will lead to self variables for other classes than the one with the current contract to prove in the namespace.
So if you try to do a cut in KeY it would think self is of type String
(as that is the first class for which this code path is taken?).

self should get put in the namespace by
`ProblemInitializer::populateNamespaces` and I don't think other classes
need self in the namespace as you need some object to call the methodes
on and that object should become self, if that proves to be false one
could use that postfix parameter instead, to not clash with plain `self`
which should stand for an object of the class the contract to prove is
from.